Basic refactor of transport, protocol, proxy layers
* Remove `SDLAbstractProtocol`, `SDLProxyFactory` classes
* Change `SDLAbstractTransport` class to `SDLTransportType` protocol
* Add some convenience initializers to `SDLProxy` and `SDLTCPTransport`
* Remove `retryDelay` from `SDLTransportType`
