+++ SmartDeviceLink/SDLIAPTransport.h
@@ -3,15 +3,17 @@
 
 #import <ExternalAccessory/ExternalAccessory.h>
 
-#import "SDLAbstractTransport.h"
+#import "SDLTransportType.h"
 #import "SDLIAPSessionDelegate.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface SDLIAPTransport : SDLAbstractTransport <SDLIAPSessionDelegate>
+@interface SDLIAPTransport : NSObject <SDLTransportType, SDLIAPSessionDelegate>
 
 @property (nullable, strong, nonatomic) SDLIAPSession *controlSession;
 @property (nullable, strong, nonatomic) SDLIAPSession *session;
+
+@property (nullable, weak, nonatomic) id<SDLTransportDelegate> delegate;
 
 @end
