+++ SmartDeviceLink/SDLProtocol.h
@@ -1,9 +1,17 @@
 //  SDLSmartDeviceLinkProtocol.h
 //
 
-#import "SDLAbstractProtocol.h"
+#import <Foundation/Foundation.h>
+
+#import "SDLTransportType.h"
+#import "SDLProtocolConstants.h"
+#import "SDLProtocolListener.h"
+#import "SDLSecurityType.h"
+#import "SDLTransportDelegate.h"
+
 @class SDLProtocolHeader;
 @class SDLProtocolRecievedMessageRouter;
+@class SDLRPCMessage;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -14,7 +22,13 @@ typedef NS_ENUM(NSUInteger, SDLProtocolError) {
 extern NSString *const SDLProtocolSecurityErrorDomain;
 
 
-@interface SDLProtocol : SDLAbstractProtocol <SDLProtocolListener>
+@interface SDLProtocol : NSObject <SDLProtocolListener, SDLTransportDelegate>
+
+@property (strong, nonatomic) NSString *debugConsoleGroupName;
+@property (nullable, weak, nonatomic) id<SDLTransportType> transport;
+@property (nullable, strong, nonatomic) NSHashTable<id<SDLProtocolListener>> *protocolDelegateTable;
+@property (nullable, nonatomic, strong) id<SDLSecurityType> securityManager;
+@property (nonatomic, copy) NSString *appId;
 
 // Sending
 - (void)startServiceWithType:(SDLServiceType)serviceType payload:(nullable NSData *)payload;
