+++ SmartDeviceLink/SDLProxy.m
@@ -6,7 +6,7 @@
 #import <UIKit/UIKit.h>
 #import <objc/runtime.h>
 
-#import "SDLAbstractTransport.h"
+#import "SDLTransportType.h"
 #import "SDLAudioStreamingState.h"
 #import "SDLLogMacros.h"
 #import "SDLEncodedSyncPData.h"
@@ -14,10 +14,10 @@
 #import "SDLFunctionID.h"
 #import "SDLGlobals.h"
 #import "SDLHMILevel.h"
+#import "SDLIAPTransport.h"
 #import "SDLLanguage.h"
 #import "SDLLayoutMode.h"
 #import "SDLLockScreenStatusManager.h"
-
 #import "SDLOnHMIStatus.h"
 #import "SDLOnSystemRequest.h"
 #import "SDLPolicyDataParser.h"
@@ -31,6 +31,7 @@
 #import "SDLStreamingMediaManager.h"
 #import "SDLSystemContext.h"
 #import "SDLSystemRequest.h"
+#import "SDLTCPTransport.h"
 #import "SDLTimer.h"
 #import "SDLVehicleType.h"
 
@@ -64,18 +65,18 @@ static float DefaultConnectionTimeout = 45.0;
 @implementation SDLProxy
 
 #pragma mark - Object lifecycle
-- (instancetype)initWithTransport:(SDLAbstractTransport *)transport protocol:(SDLAbstractProtocol *)protocol delegate:(NSObject<SDLProxyListener> *)theDelegate {
+- (instancetype)initWithTransport:(id<SDLTransportType>)transport delegate:(id<SDLProxyListener>)delegate {
     if (self = [super init]) {
         SDLLogD(@"Framework Version: %@", self.proxyVersion);
-        _debugConsoleGroupName = @"default";
         _lsm = [[SDLLockScreenStatusManager alloc] init];
         _rpcProcessingQueue = dispatch_queue_create("com.sdl.rpcProcessingQueue", DISPATCH_QUEUE_SERIAL);
-
-        _mutableProxyListeners = [NSMutableSet setWithObject:theDelegate];
+        _mutableProxyListeners = [NSMutableSet setWithObject:delegate];
         _securityManagers = [NSMutableDictionary dictionary];
-        _protocol = protocol;
+
+        _protocol = [[SDLProtocol alloc] init];
         _transport = transport;
-        _transport.delegate = protocol;
+        _transport.delegate = _protocol;
+
         [_protocol.protocolDelegateTable addObject:self];
         _protocol.transport = transport;
 
@@ -94,6 +95,21 @@ static float DefaultConnectionTimeout = 45.0;
     }
 
     return self;
+}
+
++ (SDLProxy *)iapProxyWithListener:(id<SDLProxyListener>)delegate {
+    SDLIAPTransport *transport = [[SDLIAPTransport alloc] init];
+    SDLProxy *ret = [[SDLProxy alloc] initWithTransport:transport delegate:delegate];
+
+    return ret;
+}
+
++ (SDLProxy *)tcpProxyWithListener:(id<SDLProxyListener>)delegate tcpIPAddress:(NSString *)ipaddress tcpPort:(NSString *)port {
+    SDLTCPTransport *transport = [[SDLTCPTransport alloc] initWithHostName:ipaddress portNumber:port];
+
+    SDLProxy *ret = [[SDLProxy alloc] initWithTransport:transport delegate:delegate];
+
+    return ret;
 }
 
 - (void)dealloc {
