+++ SmartDeviceLink/SDLStreamingMediaLifecycleManager.h
@@ -15,10 +15,10 @@
 #import "SDLStreamingAudioManagerType.h"
 #import "SDLStreamingMediaManagerConstants.h"
 
-@class SDLAbstractProtocol;
 @class SDLAudioStreamManager;
 @class SDLCarWindow;
 @class SDLImageResolution;
+@class SDLProtocol;
 @class SDLStateMachine;
 @class SDLStreamingMediaConfiguration;
 @class SDLTouchManager;
@@ -177,7 +177,7 @@ extern SDLAudioStreamState *const SDLAudioStreamStateShuttingDown;
 /**
  *  Start the manager with a completion block that will be called when startup completes. This is used internally. To use an SDLStreamingMediaManager, you should use the manager found on `SDLManager`.
  */
-- (void)startWithProtocol:(SDLAbstractProtocol *)protocol;
+- (void)startWithProtocol:(SDLProtocol *)protocol;
 
 /**
  *  Stop the manager. This method is used internally.
