+++ SmartDeviceLink/SDLTCPTransport.h
@@ -1,16 +1,19 @@
 //  SDLTCPTransport.h
 //
 
-#import "SDLAbstractTransport.h"
+#import "SDLTransportType.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface SDLTCPTransport : SDLAbstractTransport {
+@interface SDLTCPTransport : NSObject <SDLTransportType> {
     _Nullable CFSocketRef socket;
 }
 
+- (instancetype)initWithHostName:(NSString *)hostName portNumber:(NSString *)portNumber;
+
 @property (strong, nonatomic) NSString *hostName;
 @property (strong, nonatomic) NSString *portNumber;
+@property (nullable, weak, nonatomic) id<SDLTransportDelegate> delegate;
 
 @end
